phoebe_prop_update
