+++ src/components/home/BlurbHome.tsx
@@ -1,21 +1,20 @@
-import Image from "next/image";
+import Cupcakes from "@/public/home/cupcakes.webp";
+import BlurbHomeProps from "./BlurbHomeProps";
 
-interface CardProps {
-  title: string;
-  description: string;
-  image: string;
-}
-const BlurbHome: React.FC<CardProps> = ({ title, description, image }) => {
+const BlurbHome = () => {
   return (
     <div>
-      <div className="flex w-[30vw] flex-col items-center">
-        <Image src={image} alt="" />
-      </div>
-      <h2 className="my-[1vh] text-[1.7vw] font-medium">{title}</h2>
-      <p className="w-[18vw] text-center text-[0.9vw] font-light leading-tight">
-        {description}
-      </p>
+      <BlurbHomeProps
+        title="Collaborations"
+        description="We collaborate with various organizations, including SAGE and St. John
+        Ambulance (SJA) Canada. We plan to host campus-wide events and are
+        actively networking with other groups focused on cardiovascular disease
+        initiatives!"
+        image={Cupcakes}
+        alt="12 Cupcakes on the table with two people sitting behind it"
+      />
     </div>
   );
 };
+
 export default BlurbHome;
